# services/api/generated/runtime_artifacts/openapi/openapi.runtime.vnext.json
{
...
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryMigrateBody": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/HistoryMigrateEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "title": "HistoryMigrateBody",
        "type": "object"
      },
...
      "JudgeGrowthEventBody": {
        "properties": {
          "details": {
            "title": "Details",
            "type": "object"
          },
          "game": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game"
          },
          "metric_type": {
            "title": "Metric Type",
            "type": "string"
          },
          "metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value"
          }
        },
        "required": [
          "metric_type"
        ],
        "title": "JudgeGrowthEventBody",
        "type": "object"
      },
...
      "SavedAnswerBody": {
        "properties": {
          "full_response": {
            "title": "Full Response",
            "type": "object"
          },
          "game_slug": {
            "title": "Game Slug",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "sources": {
            "items": {
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "game_slug",
          "question"
        ],
        "title": "SavedAnswerBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "TCG Judge API",
    "version": "0.1.0-mvp"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runtime/judge/account": {
      "delete": {
        "description": "LGPD: remove dados do utilizador nas tabelas Judge (sess\u00f5es, favoritos, m\u00e9tricas).",
        "operationId": "delete_judge_account_runtime_judge_account_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Judge Account Runtime Judge Account Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Judge Account",
        "tags": [
          "judge-product"
        ]
      }
    },
...
    "/runtime/judge/growth": {
      "get": {
        "operationId": "get_judge_growth_runtime_judge_growth_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Judge Growth Runtime Judge Growth Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Judge Growth",
        "tags": [
          "judge-product"
        ]
      }
    },
    "/runtime/judge/growth/event": {
      "post": {
        "operationId": "post_growth_event_runtime_judge_growth_event_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeGrowthEventBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Post Growth Event Runtime Judge Growth Event Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Growth Event",
        "tags": [
          "judge-product"
        ]
      }
    },
...
    "/runtime/judge/history/migrate": {
      "post": {
        "description": "Migra hist\u00f3rico localStorage \u2192 sess\u00f5es cloud (idempotente por t\u00edtulo+pergunta).",
        "operationId": "migrate_judge_history_runtime_judge_history_migrate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryMigrateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Migrate Judge History Runtime Judge History Migrate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Migrate Judge History",
        "tags": [
          "judge-product"
        ]
      }
    },
...
    "/runtime/judge/saved-answers": {
      "get": {
        "operationId": "list_saved_answers_runtime_judge_saved_answers_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Saved Answers Runtime Judge Saved Answers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Saved Answers",
        "tags": [
          "judge-product"
        ]
      },
      "post": {
        "operationId": "create_saved_answer_runtime_judge_saved_answers_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedAnswerBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Saved Answer Runtime Judge Saved Answers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Saved Answer",
        "tags": [
          "judge-product"
        ]
      }
    },
    "/runtime/judge/session": {
      "post": {
        "operationId": "create_judge_session_runtime_judge_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeSessionCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Judge Session",
        "tags": [
          "judge-product"
        ]
      }
    },
    "/runtime/judge/session/{session_id}": {
      "get": {
        "operationId": "get_judge_session_runtime_judge_session__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Judge-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Judge-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgeSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Judge Session",
        "tags": [
          "judge-product"
        ]
      }
    },
    "/runtime/judge/share": {
      "post": {
        "operationId": "create_judge_share_runtime_judge_share_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JudgeShareCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgeShareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Judge Share",
        "tags": [
          "judge-product"
        ]
      }
    },
    "/runtime/judge/share/{share_id}": {
      "get": {
        "operationId": "get_judge_share_runtime_judge_share__share_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "share_id",
            "required": true,
            "schema": {
              "title": "Share Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sig",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sig"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JudgeShareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Judge Share",
        "tags": [
          "judge-product"
        ]
      }
    },
...
  }
}